+++ tests/tests.php
@@ -14,7 +14,7 @@ include_once '../ViewBuilder.php';
 
 use Exacodis\Pilot;
 
-$pilot = new Pilot('PhpEcho - A native PHP template engine - v.6.1.0');
+$pilot = new Pilot('PhpEcho - A native PHP template engine - v.6.1.1');
 $pilot->injectStandardHelpers();
 
 include 'filepath.php';